Move FlowProgressIndicator from outer signup area to inner wrappers for easier targeting and styling; since we're moving to a two-column layout, it needs to be accessible within the grid area.
