+++ client/signup/steps/about/style.scss
@@ -168,6 +168,7 @@
 		grid-column-start: 1;
 		grid-row-start: 2;
 		align-self: end;
+		margin-top: 84px;
 	}
 
 	.about__section .formatted-header {
